+++ NEWS.md
@@ -11,9 +11,10 @@ for demonstration purposes (#240).
 - Add helpful reference landing page at `polars.github.io/reference_home` (#223, #264).
 - rust-polars' `simd` feature is now disabled by default. To enable it, set the environment variable
   `RPOLARS_ALL_FEATURES` to `true` when build r-polars (#262).
-- A new function `polars_info()` will tell which features enabled (#271).
+- A new function `polars_info()` will tell which features enabled (#271, #285).
 - `select()` now accepts lists of expressions. For example, `<DataFrame>$select(l_expr)`
   works with `l_expr = list(pl$col("a"))` (#265).
+- `<DataFrame>$glimpse()` is a fast `str()`-like view of a `DataFrame` (#277).
 
 # polars 0.6.1
 ## What's changed